+++ changelog.md
@@ -4,3 +4,6 @@ a release
 -->
 
 ## Notable Changes
+
+- **Addition**: Add `StoreFileRef::encryption_keys()` function to retrieve the keys used for encrypting the given
+    file.